Chore: Moves Card component into folder as we'll have some child components
