+++ src/components/Grid.js
@@ -1,4 +1,4 @@
-import Card from './Card';
+import Card from './Card/Card';
 
 const Grid = ({ array }) => {
   return (
